Add minor changes Model Loader

+++ Chaos-Engine/Model.h
@@ -73,7 +73,7 @@ private:
 		std::vector<glm::vec2>& in_uvs,
 		std::vector<glm::vec3>& in_normals,
 
-		std::vector<unsigned short>& out_indices,
+		std::vector<GLuint>& out_indices,
 		std::vector<glm::vec3>& out_vertices,
 		std::vector<glm::vec2>& out_uvs,
 		std::vector<glm::vec3>& out_normals
